Изменена длина всех полей фио и логина на 200 символов. Изменён тип данных процентов.

+++ src/main/java/cinimex/org/entity/CreditEntity.java
@@ -3,6 +3,7 @@ package cinimex.org.entity;
 import lombok.Data;
 
 import javax.persistence.*;
+import java.math.BigDecimal;
 import java.sql.Timestamp;
 
 @Entity
@@ -20,7 +21,7 @@ public class CreditEntity {
     private Integer amount;
 
     @Column(name = "annual_rate")
-    private Integer annualRate;
+    private BigDecimal annualRate;
 
     @Column(name = "date_of_issue")
     private Timestamp dateOfIssue;
